pyproject.toml
[project]
name = "voxflowai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiolimiter>=1.2.1",
    "anthropic>=0.79.0",
    "elevenlabs>=2.34.0",
    "fastapi>=0.128.2",
    "feedparser>=6.0.12",
    "gtts>=2.5.4",
    "langchain-anthropic>=1.3.2",
    "langchain-groq>=1.1.2",
    "langchain-mcp-adapters>=0.2.1",
    "langgraph>=1.0.8",
    "lxml>=6.0.2",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
